+++ packages/react-form-composer/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 No unreleased changes for you to be aware of.
 
+## [3.1.0] - 2019-09-11
+### Added
+- Reinstate mount and unmount for Form. The Field spy prop can now be a function or string.
+- Add formRef and apiRef props to form.
+
 ## [3.0.1] - 2019-09-10
 ### Fixed
 - deregister fields so the errors are decrement when a dynamic field unmounts